Refactor: extract functions to different files; add more unit tests, achieving 100% coverage.
