+++ merkle_tree.go
@@ -30,8 +30,6 @@ import (
 	"math/bits"
 	"runtime"
 	"sync"
-
-	"golang.org/x/sync/errgroup"
 )
 
 const (
@@ -99,12 +97,6 @@ type MerkleTree struct {
 	// NumLeaves is the number of leaves in the Merkle Tree.
 	// This value is fixed once the tree is built.
 	NumLeaves int
-}
-
-// Proof represents a Merkle Tree proof.
-type Proof struct {
-	Siblings [][]byte // Sibling nodes to the Merkle Tree path of the data block.
-	Path     uint32   // Path variable indicating whether the neighbor is on the left or right.
 }
 
 // New generates a new Merkle Tree with the specified configuration and data blocks.
@@ -182,17 +174,17 @@ func New(config *Config, blocks []DataBlock) (m *MerkleTree, err error) {
 	// Build the tree in ModeTreeBuild.
 	if m.Mode == ModeTreeBuild {
 		if m.RunInParallel {
-			err = m.buildTreeParallel()
+			err = m.treeBuildParallel()
 			return
 		}
-		err = m.buildTree()
+		err = m.treeBuild()
 		return
 	}
 
 	// Build the tree and generate proofs in ModeProofGenAndTreeBuild.
 	if m.Mode == ModeProofGenAndTreeBuild {
 		if m.RunInParallel {
-			if err = m.buildTreeParallel(); err != nil {
+			if err = m.treeBuildParallel(); err != nil {
 				return
 			}
 			m.initProofs()
@@ -201,7 +193,7 @@ func New(config *Config, blocks []DataBlock) (m *MerkleTree, err error) {
 			}
 			return
 		}
-		if err = m.buildTree(); err != nil {
+		if err = m.treeBuild(); err != nil {
 			return
 		}
 		m.initProofs()
@@ -232,233 +224,4 @@ func concatSortHash(b1 []byte, b2 []byte) []byte {
 		return concatHash(b1, b2)
 	}
 	return concatHash(b2, b1)
-}
-
-// buildTree builds the Merkle Tree and stores all the nodes.
-func (m *MerkleTree) buildTree() (err error) {
-	finishMap := make(chan struct{})
-	go m.routineBuildLeafMap(finishMap)
-	m.initNodes()
-	for i := 0; i < m.Depth-1; i++ {
-		m.nodes[i] = appendNodeIfOdd(m.nodes[i])
-		numNodes := len(m.nodes[i])
-		m.nodes[i+1] = make([][]byte, numNodes>>1)
-		for j := 0; j < numNodes; j += 2 {
-			if m.nodes[i+1][j>>1], err = m.HashFunc(
-				m.concatHashFunc(m.nodes[i][j], m.nodes[i][j+1]),
-			); err != nil {
-				return
-			}
-		}
-	}
-	if m.Root, err = m.HashFunc(m.concatHashFunc(
-		m.nodes[m.Depth-1][0], m.nodes[m.Depth-1][1],
-	)); err != nil {
-		return
-	}
-	<-finishMap
-	return
-}
-
-func (m *MerkleTree) routineBuildLeafMap(finishChan chan struct{}) {
-	m.leafMapMu.Lock()
-	defer m.leafMapMu.Unlock()
-	for i := 0; i < m.NumLeaves; i++ {
-		m.leafMap[string(m.Leaves[i])] = i
-	}
-	finishChan <- struct{}{} // empty channel to serve as a wait group for map generation
-}
-
-func appendNodeIfOdd(buffer [][]byte) [][]byte {
-	if len(buffer)&1 == 0 {
-		return buffer
-	}
-	appendNode := buffer[len(buffer)-1]
-	buffer = append(buffer, appendNode)
-	return buffer
-}
-
-func (m *MerkleTree) initNodes() {
-	m.nodes = make([][][]byte, m.Depth)
-	m.nodes[0] = make([][]byte, m.NumLeaves)
-	copy(m.nodes[0], m.Leaves)
-}
-
-// buildTreeParallel builds the Merkle Tree and stores all the nodes in parallel.
-func (m *MerkleTree) buildTreeParallel() (err error) {
-	finishMap := make(chan struct{})
-	go m.routineBuildLeafMap(finishMap)
-	m.initNodes()
-	for i := 0; i < m.Depth-1; i++ {
-		m.nodes[i] = appendNodeIfOdd(m.nodes[i])
-		numNodes := len(m.nodes[i])
-		m.nodes[i+1] = make([][]byte, numNodes>>1)
-		numRoutines := m.NumRoutines
-		if numRoutines > numNodes {
-			numRoutines = numNodes
-		}
-		eg := new(errgroup.Group)
-		for startIdx := 0; startIdx < numRoutines; startIdx++ {
-			startIdx := startIdx
-			eg.Go(func() error {
-				for j := startIdx << 1; j < numNodes; j += numRoutines << 1 {
-					newHash, err := m.HashFunc(m.concatHashFunc(
-						m.nodes[i][j], m.nodes[i][j+1],
-					))
-					if err != nil {
-						return err
-					}
-					m.nodes[i+1][j>>1] = newHash
-				}
-				return nil
-			})
-		}
-		if err = eg.Wait(); err != nil {
-			return
-		}
-	}
-	if m.Root, err = m.HashFunc(m.concatHashFunc(
-		m.nodes[m.Depth-1][0], m.nodes[m.Depth-1][1],
-	)); err != nil {
-		return
-	}
-	<-finishMap
-	return
-}
-
-func (m *MerkleTree) buildProofsFromNodes(buffer [][]byte, bufferLength, step int) {
-	batch := 1 << step
-	for i := 0; i < bufferLength; i += 2 {
-		m.buildProofPairsFromNodes(buffer, i, batch, step)
-	}
-}
-
-func (m *MerkleTree) buildProofsFromNodesParallel(buffer [][]byte, bufferLength, step int) {
-	batch := 1 << step
-	numRoutines := m.NumRoutines
-	if numRoutines > bufferLength {
-		numRoutines = bufferLength
-	}
-	var wg sync.WaitGroup
-	wg.Add(numRoutines)
-	for startIdx := 0; startIdx < numRoutines; startIdx++ {
-		go func(startIdx int) {
-			defer wg.Done()
-			for i := startIdx; i < bufferLength; i += numRoutines << 1 {
-				m.buildProofPairsFromNodes(buffer, i, batch, step)
-			}
-		}(startIdx << 1)
-	}
-	wg.Wait()
-}
-
-func (m *MerkleTree) buildProofPairsFromNodes(buffer [][]byte, idx, batch, step int) {
-	start := idx * batch
-	end := min(start+batch, len(m.Proofs))
-	for i := start; i < end; i++ {
-		m.Proofs[i].Path += 1 << step
-		m.Proofs[i].Siblings = append(m.Proofs[i].Siblings, buffer[idx+1])
-	}
-	start += batch
-	end = min(start+batch, len(m.Proofs))
-	for i := start; i < end; i++ {
-		m.Proofs[i].Siblings = append(m.Proofs[i].Siblings, buffer[idx])
-	}
-}
-
-// Verify checks if the data block is valid using the Merkle Tree proof and the cached Merkle root hash.
-func (m *MerkleTree) Verify(dataBlock DataBlock, proof *Proof) (bool, error) {
-	return Verify(dataBlock, proof, m.Root, m.Config)
-}
-
-// Verify checks if the data block is valid using the Merkle Tree proof and the provided Merkle root hash.
-// It returns true if the data block is valid, false otherwise. An error is returned in case of any issues
-// during the verification process.
-func Verify(dataBlock DataBlock, proof *Proof, root []byte, config *Config) (bool, error) {
-	// Validate input parameters.
-	if dataBlock == nil {
-		return false, ErrDataBlockIsNil
-	}
-	if proof == nil {
-		return false, ErrProofIsNil
-	}
-	if config == nil {
-		config = new(Config)
-	}
-	if config.HashFunc == nil {
-		config.HashFunc = DefaultHashFunc
-	}
-
-	// Determine the concatenation function based on the configuration.
-	concatFunc := concatHash
-	if config.SortSiblingPairs {
-		concatFunc = concatSortHash
-	}
-
-	// Convert the data block to a leaf.
-	leaf, err := dataBlockToLeaf(dataBlock, config.HashFunc, config.DisableLeafHashing)
-	if err != nil {
-		return false, err
-	}
-
-	// Traverse the Merkle proof and compute the resulting hash.
-	// Copy the slice so that the original leaf won't be modified.
-	result := make([]byte, len(leaf))
-	copy(result, leaf)
-	path := proof.Path
-	for _, sib := range proof.Siblings {
-		if path&1 == 1 {
-			result, err = config.HashFunc(concatFunc(result, sib))
-		} else {
-			result, err = config.HashFunc(concatFunc(sib, result))
-		}
-		if err != nil {
-			return false, err
-		}
-		path >>= 1
-	}
-	return bytes.Equal(result, root), nil
-}
-
-// Proof generates the Merkle proof for a data block using the previously generated Merkle Tree structure.
-// This method is only available when the configuration mode is ModeTreeBuild or ModeProofGenAndTreeBuild.
-// In ModeProofGen, proofs for all the data blocks are already generated, and the Merkle Tree structure
-// is not cached.
-func (m *MerkleTree) Proof(dataBlock DataBlock) (*Proof, error) {
-	if m.Mode != ModeTreeBuild && m.Mode != ModeProofGenAndTreeBuild {
-		return nil, ErrProofInvalidModeTreeNotBuilt
-	}
-
-	// Convert the data block to a leaf.
-	leaf, err := dataBlockToLeaf(dataBlock, m.HashFunc, m.DisableLeafHashing)
-	if err != nil {
-		return nil, err
-	}
-
-	// Retrieve the index of the leaf in the Merkle Tree.
-	m.leafMapMu.Lock()
-	idx, ok := m.leafMap[string(leaf)]
-	m.leafMapMu.Unlock()
-	if !ok {
-		return nil, ErrProofInvalidDataBlock
-	}
-
-	// Compute the path and siblings for the proof.
-	var (
-		path     uint32
-		siblings = make([][]byte, m.Depth)
-	)
-	for i := 0; i < m.Depth; i++ {
-		if idx&1 == 1 {
-			siblings[i] = m.nodes[i][idx-1]
-		} else {
-			path += 1 << i
-			siblings[i] = m.nodes[i][idx+1]
-		}
-		idx >>= 1
-	}
-	return &Proof{
-		Path:     path,
-		Siblings: siblings,
-	}, nil
 }
