+++ proof_gen.go
@@ -1,3 +1,28 @@
+// MIT License
+//
+// Copyright (c) 2023 Tommy TIAN
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+// Package merkletree implements a high-performance Merkle Tree in Go.
+// It supports parallel execution for enhanced performance and
+// offers compatibility with OpenZeppelin through sorted sibling pairs.
 package merkletree
 
 import (
@@ -35,14 +60,14 @@ func (m *MerkleTree) generateProofsParallel() (err error) {
 	numRoutines := m.NumRoutines
 	for step := 0; step < m.Depth; step++ {
 		// Limit the number of workers to the previous level length.
-		if numRoutines > bufferSize {
-			numRoutines = bufferSize
-		}
+		numRoutines = min(numRoutines, bufferSize)
 		bufferSize = fixOddNumOfNodes(buffer, bufferSize, step)
 		m.updateProofsParallel(buffer, bufferSize, step)
-		eg := new(errgroup.Group)
-		hashFunc := m.HashFunc
-		concatHashFunc := m.concatHashFunc
+		var (
+			eg             = new(errgroup.Group)
+			hashFunc       = m.HashFunc
+			concatHashFunc = m.concatHashFunc
+		)
 		for startIdx := 0; startIdx < numRoutines; startIdx++ {
 			startIdx := startIdx << 1
 			eg.Go(func() error {
@@ -120,24 +145,23 @@ func fixOddNumOfNodes(buffer [][]byte, bufferSize, step int) int {
 func (m *MerkleTree) updateProofs(buffer [][]byte, bufferSize, step int) {
 	batch := 1 << step
 	for i := 0; i < bufferSize; i += 2 {
-		m.updateProofPairs(buffer, i, batch, step)
+		updateProofPairs(m.Proofs, buffer, i, batch, step)
 	}
 }
 
 // updateProofsParallel updates the proofs for all the leaves while constructing the Merkle Tree in parallel.
 func (m *MerkleTree) updateProofsParallel(buffer [][]byte, bufferLength, step int) {
-	batch := 1 << step
-	numRoutines := m.NumRoutines
-	if numRoutines > bufferLength {
-		numRoutines = bufferLength
-	}
-	var wg sync.WaitGroup
+	var (
+		batch       = 1 << step
+		numRoutines = min(m.NumRoutines, bufferLength)
+		wg          sync.WaitGroup
+	)
 	wg.Add(numRoutines)
 	for startIdx := 0; startIdx < numRoutines; startIdx++ {
 		go func(startIdx int) {
 			defer wg.Done()
 			for i := startIdx; i < bufferLength; i += numRoutines << 1 {
-				m.updateProofPairs(buffer, i, batch, step)
+				updateProofPairs(m.Proofs, buffer, i, batch, step)
 			}
 		}(startIdx << 1)
 	}
@@ -145,18 +169,18 @@ func (m *MerkleTree) updateProofsParallel(buffer [][]byte, bufferLength, step in
 }
 
 // updateProofPairs updates the proofs in the Merkle Tree in pairs.
-func (m *MerkleTree) updateProofPairs(buffer [][]byte, idx, batch, step int) {
+func updateProofPairs(proofs []*Proof, buffer [][]byte, idx, batch, step int) {
 	start := idx * batch
-	end := min(start+batch, len(m.Proofs))
+	end := min(start+batch, len(proofs))
 	siblingNodeIdx := min((idx+1)<<step, len(buffer)-1)
 	for i := start; i < end; i++ {
-		m.Proofs[i].Path += 1 << step
-		m.Proofs[i].Siblings = append(m.Proofs[i].Siblings, buffer[siblingNodeIdx])
+		proofs[i].Path += 1 << step
+		proofs[i].Siblings = append(proofs[i].Siblings, buffer[siblingNodeIdx])
 	}
 	start += batch
-	end = min(start+batch, len(m.Proofs))
+	end = min(start+batch, len(proofs))
 	siblingNodeIdx = min(idx<<step, len(buffer)-1)
 	for i := start; i < end; i++ {
-		m.Proofs[i].Siblings = append(m.Proofs[i].Siblings, buffer[siblingNodeIdx])
+		proofs[i].Siblings = append(proofs[i].Siblings, buffer[siblingNodeIdx])
 	}
 }
